Update bar on wallpaper change
